+++ lib/src/widget_extensions.dart
@@ -41,16 +41,16 @@ extension WidgetExtensions on Widget {
 
   /// Adds uniform margin to all sides using a wrapping [Container].
   @Deprecated(
-    'Deprecated in v2.0. Use `paddingAll` on the parent or a `Spacing` widget '
-    'in the sibling list instead. Will be removed in v3.0.',
+    'Deprecated in v1.0. Use `paddingAll` on the parent or a `Spacing` widget '
+    'in the sibling list instead. Will be removed in v2.0.',
   )
   Widget marginAll(double value) =>
       Container(margin: EdgeInsets.all(value), child: this);
 
   /// Adds symmetric horizontal and vertical margin using a wrapping [Container].
   @Deprecated(
-    'Deprecated in v2.0. Use `paddingSymmetric` or a `Spacing` widget instead. '
-    'Will be removed in v3.0.',
+    'Deprecated in v1.0. Use `paddingSymmetric` or a `Spacing` widget instead. '
+    'Will be removed in v2.0.',
   )
   Widget marginSymmetric({double horizontal = 0, double vertical = 0}) =>
       Container(
@@ -61,8 +61,8 @@ extension WidgetExtensions on Widget {
 
   /// Adds directional margin (left, top, right, bottom) using a wrapping [Container].
   @Deprecated(
-    'Deprecated in v2.0. Use `paddingOnly` or a `Spacing` widget instead. '
-    'Will be removed in v3.0.',
+    'Deprecated in v1.0. Use `paddingOnly` or a `Spacing` widget instead. '
+    'Will be removed in v2.0.',
   )
   Widget marginOnly({
     double left = 0,
@@ -124,7 +124,9 @@ extension WidgetExtensions on Widget {
   /// ```dart
   /// Text('Tap me').onTap(() => print('Tapped!'));
   /// ```
-  Widget onTap(VoidCallback onTap,
-          {HitTestBehavior behavior = HitTestBehavior.opaque}) =>
+  Widget onTap(
+    VoidCallback onTap, {
+    HitTestBehavior behavior = HitTestBehavior.opaque,
+  }) =>
       GestureDetector(onTap: onTap, behavior: behavior, child: this);
 }